% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_prop.R
\name{count_prop}
\alias{count_prop}
\title{Get grouped count and proportion}
\usage{
count_prop(x, ..., sort = FALSE)
}
\arguments{
\item{x}{A data frame, data frame extension (tibble), or a lazy data frame.}

\item{...}{Variables to group by.}

\item{sort}{If ‘TRUE’, will show the largest groups at the top.}
}
\value{
An object of the same type as `.data`
}
\description{
Similar to `count`, but also includes a `prop` column with the
proportion for each group.
}
\examples{
library(dplyr)
starwars |> count_prop(species)
}
