% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{is the angle in degrees}
}
\value{
returns the angle in radians
}
\description{
Convert degrees to radians
}
\examples{
rad2deg(3.1415)
}
