% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gz.R
\name{gz}
\alias{gz}
\title{Gzip Files}
\usage{
gz(src, dest)
}
\arguments{
\item{src}{filesystem path of the file to be gzipped}

\item{dest}{filesystem path of the resulting gzipped file}
}
\value{
None
}
\description{
Helper to gzip files on the filesystem.
}
\examples{
\dontrun{
gz("plot.svg", "plot.svgz")
}
}
