% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_path_original.R
\name{link_path_original}
\alias{link_path_original}
\title{Trace a symlink back to it's source file}
\usage{
link_path_original(symlink, count = 0)
}
\arguments{
\item{symlink}{Symbolic link to trace back}

\item{count}{Number of recursive calls}
}
\value{
path to original file
}
\description{
Trace a symlink back to it's source file
}
\examples{
library(fs)
file_create("/tmp/a")
link_create("/tmp/a", "/tmp/b")
link_create("/tmp/b", "/tmp/c")
link_path_original("/tmp/c")
}
