% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_to_hhmmss.R
\name{min_to_hhmmss}
\alias{min_to_hhmmss}
\title{Convert minutes to HH:MM:SS.s string}
\usage{
min_to_hhmmss(minutes)
}
\arguments{
\item{minutes}{numeric for the minutes to format}
}
\value{
character string formatted as HH:MM:SS.s
}
\description{
Convert minutes into a zero padded string formatted as HH:MM:SS.s.
}
\examples{
min_to_hhmmss(139.56)
}
