% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_to_mmss.R
\name{min_to_mmss}
\alias{min_to_mmss}
\title{Convert minutes to MM:SS.s string}
\usage{
min_to_mmss(minutes)
}
\arguments{
\item{minutes}{numeric for the minutes to format}
}
\value{
character string formatted as MM:SS.s
}
\description{
Convert minutes into a zero padded string formatted as MM:SS.s.
}
\examples{
min_to_mmss(9.56)
}
