% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{Find the percentile of an observation in a list of observations.}
\usage{
percentile(value, x, from = 0, to = 1, eps = 0.000001)
}
\arguments{
\item{value}{numeric. A scalar value to compare against the observations.}

\item{x}{numeric. A numeric vector.}

\item{from}{numeric. Start interval for internal logic.}

\item{to}{numeric. End interval for internal logic.}

\item{eps}{numeric. Threshold for calculating approximate value.}
}
\value{
numeric. Percentile position of value between [0, 100].
}
\description{
Based on https://github.com/choonghyunryu/dlookr/blob/master/R/utils.R
}
\examples{
x <- runif(100) * 100
percentile(25, x)
}
