% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_wkb_as_sf.R
\name{pg_wkb_as_sf}
\alias{pg_wkb_as_sf}
\title{Convert a data frame with a PostGIS WKB column to a sf object.}
\usage{
pg_wkb_as_sf(x, geometry)
}
\arguments{
\item{x}{data frame containing a PostGIS WKB column}

\item{geometry}{name of the geometry column}
}
\value{
sf object
}
\description{
Convert a data frame with a PostGIS WKB column to a sf object.
}
\examples{
\dontrun{
library(RPostgres)
library(sf)
con <- dbConnect(
  Postgres(),
  host = "localhost",
  dbname = "gis_db"
)
df <- con |>
  tbl("gis_data") |>
  collect()
sf_df <- df |>
  pg_wkb_as_sf("the_geom")
}
}
