% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{is the angle in radians}
}
\value{
returns the angle in degrees
}
\description{
Convert radians to degrees
}
\examples{
rad2deg(3.1415)
}
