% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_width.R
\name{reduce_width}
\alias{reduce_width}
\title{Reduces the option "width" by two}
\usage{
reduce_width()
}
\description{
Reduces the width by two from the current width
}
\examples{
getOption("width")
reduce_width()
getOption("width")
}
