% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_geometry.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename geometry column}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{spatial data frame containing an sf geometry}

\item{name}{updated geometry name}
}
\value{
spatial data frame with the geometry column renamed
}
\description{
Rename the geometry column in an sf object
}
\details{
Function code from Spacedman in this thread:·https://gis.stackexchange.com/questions/386589
}
\examples{
library(sf)
df <- data.frame(name = "ABR", lat = 64.912, lon = -147.932)
df <- st_as_sf(df, coords = c("lon", "lat"), crs = 4326)
df <- rename_geometry(df, "the_geom")
}
