% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Add coordinates as new columns}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{x}{data frame containing an sf point geometry}

\item{names}{character vector containing the names of the new columns to add}
}
\value{
data frame with the new columns added
}
\description{
Extract the coordinates from the data frame and create two new columns
containing these coordinates.
}
\examples{
library(sf)
df <- data.frame(name = "ABR", lat = 64.912, lon = -147.932)
df <- st_as_sf(df, coords = c("lon", "lat"), crs = 4326)
df <- sfc_as_cols(df, c("lon_wgs84", "lat_wgs84"))
}
