% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_str_agg.R
\name{unique_str_agg}
\alias{unique_str_agg}
\title{Unique values string aggregate}
\usage{
unique_str_agg(x, collapse = ",")
}
\arguments{
\item{x}{character vector to aggregate}

\item{collapse}{string placed between elements in aggregate}
}
\value{
character value with all unique items included, separated by the value of the collapse parameter.
}
\description{
String aggregate that only includes unique values.
}
\examples{
x <- c("a", "a", "b", "c", "d", "d")
unique_str_agg(x, ",")
}
