% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utmzone.R
\name{utmzone}
\alias{utmzone}
\title{Add utmzone SRID as a new column in data frame}
\usage{
utmzone(x)
}
\arguments{
\item{x}{data frame containing an sf point geometry}
}
\value{
data frame with the utmzone column added
}
\description{
Create a new utmzone column containing the EPSG code for the
UTM zone the points are within.
}
\examples{
library(sf)
df <- data.frame(name = "ABR", lat = 64.912, lon = -147.932)
df <- st_as_sf(df, coords = c("lon", "lat"), crs = 4326)
df <- utmzone(df)
}
